# swagger.yaml
openapi: 3.0.1
info:
  title: ConfSystem
  version: v1
paths:
...
  /agendas-module/CallForPapers/open:
    put:
      tags:
        - CallForPapers
      parameters:
        - name: conferenceId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
  /agendas-module/CallForPapers/close:
    put:
      tags:
        - CallForPapers
      parameters:
        - name: conferenceId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
...
  /tickets-module/Sales/conferences/{conferenceId}:
    get:
      tags:
        - Sales
      parameters:
        - name: conferenceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConfSystem.Modules.Tickets.Core.DTO.TicketSaleInfoDto'
    post:
      tags:
        - Sales
      parameters:
        - name: conferenceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfSystem.Modules.Tickets.Core.DTO.TicketSaleDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ConfSystem.Modules.Tickets.Core.DTO.TicketSaleDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ConfSystem.Modules.Tickets.Core.DTO.TicketSaleDto'
      responses:
        '200':
          description: Success
  /tickets-module/Sales/conferences/{conferenceId}/current:
    get:
      tags:
        - Sales
      parameters:
        - name: conferenceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfSystem.Modules.Tickets.Core.DTO.TicketSaleInfoDto'
  /speakers-module/Speakers/{id}:
    get:
      tags:
        - Speakers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfSystem.Modules.Speakers.Core.DTO.SpeakerDto'
    put:
      tags:
        - Speakers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfSystem.Modules.Speakers.Core.DTO.SpeakerDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ConfSystem.Modules.Speakers.Core.DTO.SpeakerDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ConfSystem.Modules.Speakers.Core.DTO.SpeakerDto'
      responses:
        '200':
          description: Success
...
  /agendas-module/Submissions/{id}:
    get:
      tags:
        - Submissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfSystem.Modules.Agendas.Application.Submissions.DTO.SubmissionDto'
...
  /agendas-module/Submissions/{id}/approve:
    put:
      tags:
        - Submissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
  /agendas-module/Submissions/{id}/reject:
    put:
      tags:
        - Submissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
...
components:
  schemas:
...
    ConfSystem.Modules.Agendas.Application.Submissions.DTO.SpeakerDto:
      type: object
      properties:
        speakerId:
          type: string
          format: uuid
        fullName:
          type: string
          nullable: true
      additionalProperties: false
    ConfSystem.Modules.Agendas.Application.Submissions.DTO.SubmissionDto:
      type: object
      properties:
        submissionId:
          type: string
          format: uuid
        conferenceId:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        level:
          type: integer
          format: int32
        status:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        speakers:
          type: array
          items:
            $ref: '#/components/schemas/ConfSystem.Modules.Agendas.Application.Submissions.DTO.SpeakerDto'
          nullable: true
      additionalProperties: false
...
    ConfSystem.Modules.Speakers.Core.DTO.SpeakerDto:
      required:
        - bio
        - email
        - fullName
      type: object
      properties:
        speakerId:
          type: string
          format: uuid
        email:
          minLength: 1
          type: string
          format: email
        fullName:
          maxLength: 100
          minLength: 3
          type: string
        bio:
          maxLength: 100
          minLength: 3
          type: string
        avatarUrl:
          type: string
          nullable: true
      additionalProperties: false
    ConfSystem.Modules.Tickets.Core.DTO.ConferenceDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
      additionalProperties: false
...
    ConfSystem.Modules.Tickets.Core.DTO.TicketSaleDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conferenceId:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        price:
          maximum: 100000
          minimum: 0
          type: number
          format: double
          nullable: true
        amount:
          type: integer
          format: int32
          nullable: true
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
      additionalProperties: false
    ConfSystem.Modules.Tickets.Core.DTO.TicketSaleInfoDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        conference:
          $ref: '#/components/schemas/ConfSystem.Modules.Tickets.Core.DTO.ConferenceDto'
        ticketPrice:
          type: number
          format: double
          nullable: true
        totalTickets:
          type: integer
          format: int32
          nullable: true
        availableTickets:
          type: integer
          format: int32
          nullable: true
        saleFrom:
          type: string
          format: date-time
        saleTo:
          type: string
          format: date-time
        isFree:
          type: boolean
          readOnly: true
        unlimitedTickets:
          type: boolean
          readOnly: true
        isAvailable:
          type: boolean
          readOnly: true
      additionalProperties: false
...